11.4.1.20250218
Updated the Grafana version to 11.4.1 and VictoriaMetrics references accordingly in the Dockerfile. Also adjusted plugin names and URLs for consistency.

fix: Update Dockerfile to use non-secure curl for API requests

The Dockerfile was updated to use a non-secure curl command for API requests, ensuring compatibility with the latest changes in external sources.